+++ src/vhost_user/mod.rs
@@ -18,6 +18,7 @@
 //! Most messages that can be sent via the Unix domain socket implementing vhost-user have an
 //! equivalent ioctl to the kernel implementation.
 
+use std::fs::File;
 use std::io::Error as IOError;
 
 pub mod message;
@@ -174,6 +175,16 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 /// Result of request handler.
 pub type HandlerResult<T> = std::result::Result<T, IOError>;
+
+/// Utility function to take the first element from option of a vector of files.
+/// Returns `None` if the vector contains no file or more than one file.
+pub(crate) fn take_single_file(files: Option<Vec<File>>) -> Option<File> {
+    let mut files = files?;
+    if files.len() != 1 {
+        return None;
+    }
+    Some(files.swap_remove(0))
+}
 
 #[cfg(all(test, feature = "vhost-user-slave"))]
 mod dummy_slave;
