+++ src/vhost_user/slave_req_handler.rs
@@ -11,7 +11,7 @@ use std::sync::{Arc, Mutex};
 use super::connection::Endpoint;
 use super::message::*;
 use super::slave_fs_cache::SlaveFsCacheReq;
-use super::{Error, Result};
+use super::{take_single_file, Error, Result};
 
 /// Services provided to the master by the slave with interior mutability.
 ///
@@ -801,16 +801,6 @@ impl<S: VhostUserSlaveReqHandler> SlaveReqHandler<S> {
             .send_message_with_payload(&hdr, msg, payload, None)?;
         Ok(())
     }
-}
-
-/// Utility function to take the first element from option of a vector of files.
-/// Returns `None` if the vector contains no file or more than one file.
-pub(crate) fn take_single_file(files: Option<Vec<File>>) -> Option<File> {
-    let mut files = files?;
-    if files.len() != 1 {
-        return None;
-    }
-    Some(files.swap_remove(0))
 }
 
 impl<S: VhostUserSlaveReqHandler> AsRawFd for SlaveReqHandler<S> {
